Added comments to the WorldUserInterface
